Project/ActualProjectBackupC/requirements.txt
pygame
mysql.connector